% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataSupportRaster}
\alias{dataSupportRaster}
\title{Get data-support-measure raster surface}
\usage{
dataSupportRaster(dat, r, chc_mask)
}
\arguments{
\item{dat}{SpatVector}

\item{r}{SpatRaster}

\item{chc_mask}{SpatVector}
}
\value{
SpatRaster
}
\description{
Get data-support-measure raster surface
}
