% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genResiduals}
\alias{genResiduals}
\title{Generate hedonic model residuals}
\usage{
genResiduals(dat, model)
}
\arguments{
\item{dat}{data.frame}

\item{model}{formula}
}
\value{
data.frame
}
\description{
Generate hedonic model residuals
}
