% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nnGetDelta}
\alias{nnGetDelta}
\title{Get delta surface}
\usage{
nnGetDelta(
  b,
  a,
  model = stats::as.formula(err ~ 1),
  raster,
  mask_layer,
  nmax,
  idp
)
}
\arguments{
\item{b}{SpatVector}

\item{a}{SpatVector}

\item{model}{formula}

\item{raster}{SpatRaster}

\item{mask_layer}{SpatVector}

\item{nmax}{integer}

\item{idp}{integer}
}
\value{
SpatRaster
}
\description{
Get delta surface
}
